timex v3.4.1 Timex.Parse.Timezones.Posix

Parses POSIX-style timezones:

Format

POSIX-style timezones are of the format: local_timezone,date/time,date/time

Where date is in the Mm.n.d format, and where:

Mm (1-12) for 12 months

n (1-5) 1 for the first week and 5 for the last week in the month

d (0-6) 0 for Sunday and 6 for Saturday

Example

TZ = CST6CDT,M3.2.0/2:00:00,M11.1.0/2:00:00

This would represents a change to daylight saving time at 2:00 AM on the second Sunday in March and change back at 2:00 AM on the first Sunday in November, and keep 6 hours time offset from GMT every year. The breakdown of the string is:

CST6CDT is the timezone name

CST is the standard abbreviation

6 is the hours of time difference from GMT

CDT is the DST abbreviation

,M3 is the third month

.2 is the second occurrence of the day in the month

.0 is Sunday

/2:00:00 is the time

,M11 is the eleventh month

.1 is the first occurrence of the day in the month

.0 is Sunday

/2:00:00 is the time

Functions

parse(tz)