timex v3.4.1 Timex.Timezone

This module is used for looking up the timezone information for a given point in time, in the desired zone. Timezones are dependent not only on locale, but the date and time for which you are querying. For instance, the timezone offset from UTC for Europe/Moscow is different for March 3rd of 2015, than it was in 2013. These differences are important, and as such, all functions in this module are date/time sensitive, and where omitted, the current date/time are assumed.

In addition to lookups, this module also does conversion of datetimes from one timezone period to another, and determining the difference between a date in one timezone period and the same date/time in another timezone period.

Functions

beginning_of_day(dt)

beginning_of_day(DateTime.t()) :: DateTime.t()

Shifts the provided DateTime to the beginning of the day in it’s timezone

convert(date, tz)

convert(date :: DateTime.t(), tz :: Timex.AmbiguousTimezoneInfo.t()) ::
  Timex.AmbiguousDateTime.t() | {:error, term()}

convert(date :: DateTime.t(), tz :: Timex.TimezoneInfo.t() | String.t()) ::
  DateTime.t() | Timex.AmbiguousDateTime.t() | {:error, term()}

Convert a date to the given timezone (either TimezoneInfo or a timezone name)

diff(dt, dest)

diff(date :: DateTime.t(), tz :: Timex.TimezoneInfo.t()) ::
  integer() | {:error, term()}

Determine what offset is required to convert a date into a target timezone

end_of_day(dt)

end_of_day(DateTime.t()) :: DateTime.t()

Shifts the provided DateTime to the end of the day in it’s timezone

exists?(zone)

exists?(String.t()) :: boolean()

Determines if a given zone name exists

get(tz, datetime \\ :calendar.universal_time())

get(Timex.Types.valid_timezone(), Timex.Types.valid_datetime()) ::
  Timex.TimezoneInfo.t() | Timex.AmbiguousTimezoneInfo.t() | {:error, term()}

Gets timezone info for a given zone name and date. The date provided can either be an Erlang datetime tuple, or a DateTime struct, and if one is not provided, then the current date and time is returned.

local()

local() ::
  Timex.TimezoneInfo.t() | Timex.AmbiguousTimezoneInfo.t() | {:error, term()}

Gets the local timezone configuration for the current date and time.

local(date)

local(Timex.Types.valid_datetime()) ::
  Timex.TimezoneInfo.t() | Timex.AmbiguousTimezoneInfo.t() | {:error, term()}

Gets the local timezone configuration for the provided date and time. The provided date and time can either be an Erlang datetime tuple, or a DateTime struct.

name_of(offset)

name_of(
  Timex.Types.valid_timezone()
  | Timex.TimezoneInfo.t()
  | Timex.AmbiguousTimezoneInfo.t()
) ::
  String.t()
  | {:error, {:invalid_timezone, term()}}
  | {:error, {:no_such_zone, term()}}
  | {:error, term()}

This function takes one of the varying timezone representations:

atoms

offset integers

shortcut names (i.e. :utc, :local, “Z”, “A”)

and resolves the full name of the timezone if it’s able.

If a string is provided which isn’t recognized, it is returned untouched, only when get/2 is called will the timezone lookup fail.

resolve(tzname, datetime, utc_or_wall \\ :wall)

resolve(String.t(), non_neg_integer(), :utc | :wall) ::
  Timex.TimezoneInfo.t() | Timex.AmbiguousTimezoneInfo.t() | {:error, term()}

resolve(Timex.Types.valid_timezone(), Timex.Types.datetime(), :utc | :wall) ::
  DateTime.t() | Timex.AmbiguousDateTime.t() | {:error, term()}

Given a timezone name as a string, and a date/time in the form of the number of seconds since year zero, attempt to resolve a TimezoneInfo for that date/time. If the time is ambiguous, AmbiguousTimezoneInfo will be returned. If no result is found, an error will be returned.

If an invalid zone name is provided, an error will be returned

total_offset(timezone_info)

total_offset(std_offset, utc_offset)

tzdata_to_timezone(tzdata, zone)

tzdata_to_timezone(map(), String.t()) :: Timex.TimezoneInfo.t()