timex v3.4.1 Timex.Timezone.Local

This module is responsible for determining the timezone configuration of the local machine. It determines this from a number of sources, depending on platform, but the order of precedence is as follows:

ALL:

TZ environment variable. Ignored if nil/empty

OSX:

/etc/localtime

systemsetup -gettimezone (if admin rights are present)

UNIX:

/etc/timezone

/etc/sysconfig/clock

/etc/conf.d/clock

/etc/localtime

/usr/local/etc/localtime

Windows:

SYSTEM registry for the currently configured TimeZoneInformation

Each location is tried, and if an error is encountered, the next is attempted, until either a successful lookup is performed, or we run out of locations to check.

Types

gregorian_seconds()

gregorian_seconds() :: non_neg_integer()

Functions

lookup()

lookup() :: String.t() | {:error, term()}

Looks up the local timezone configuration. Returns the name of a timezone in the Olson database.

If no reference time is provided (in gregorian seconds), the current time in UTC will be used. If one is provided, the reference time will be used to find the local timezone for that reference time, if it exists.

lookup(secs)

lookup(gregorian_seconds()) :: String.t() | {:error, term()}

parse_tzfile(tzdata)

parse_tzfile(binary()) :: {:ok, String.t()} | {:error, term()}

Given a binary representing the data from a tzfile (not the source version), parses out the timezone for the curent date/time in UTC.

parse_tzfile(tzdata, reference_date)

parse_tzfile(binary(), gregorian_seconds()) ::
  {:ok, String.t()} | {:error, term()}

Same as parse_tzfile/1, but takes a reference date (in gregorian seconds). The reference date will be used to locate the timezone period for the local timezone which applies to that date.