timex v3.4.1 Timex.Translator

Functions

current_locale()

current_locale() :: String.t()

Returns the active locale for the process in which this function is called

default_locale()

default_locale() :: String.t()

Returns the currently configured default locale. If not set, “en” is used.

get_day_periods(locale)

get_day_periods(locale :: String.t()) :: %{optional(atom()) => String.t()}

Returns a map of day period types to translated day period names

Examples

iex> day_periods = Timex.Translator.get_day_periods("en")
...> {day_periods[:am], day_periods[:AM]}
{"am", "AM"}

get_months(locale)

get_months(locale :: String.t()) :: %{optional(integer()) => String.t()}

Returns a map of ordinal months to month names

get_months_abbreviated(locale)

get_months_abbreviated(locale :: String.t()) :: %{
  optional(integer()) => String.t()
}

Returns a map of ordinal months to month abbreviations

get_weekdays(locale)

get_weekdays(locale :: String.t()) :: %{optional(integer()) => String.t()}

Returns a map of ordinal weekdays to weekday names, where Monday = 1, translated in the given locale

get_weekdays_abbreviated(locale)

get_weekdays_abbreviated(locale :: String.t()) :: %{
  optional(integer()) => String.t()
}

Returns a map of ordinal weekdays to weekday abbreviations, where Mon = 1

translate(locale, domain, msgid)

translate(locale :: String.t(), domain :: String.t(), msgid :: String.t()) ::
  String.t()

Translates a string for a given locale and domain.

Examples

iex> Timex.Translator.translate("ru", "weekdays", "Saturday")
"суббота"

iex> Timex.Translator.translate("it", "weekdays", "Saturday")
"Sabato"

iex> Timex.Translator.translate("invalid_locale", "weekdays", "Saturday")
"Saturday"

translate_plural(locale, domain, msgid, msgid_plural, n)

translate_plural(
  locale :: String.t(),
  domain :: String.t(),
  msgid :: String.t(),
  msgid_plural :: String.t(),
  n :: integer()
) :: String.t()

Translates a string for a given locale and domain, following the pluralization rules of that language.

Examples

iex> Timex.Translator.translate_plural("ru", "relative_time", "in %{count} second", "in %{count} seconds", 5)
"через 5 секунд"

iex> Timex.Translator.translate_plural("it", "relative_time", "in %{count} second", "in %{count} seconds", 5)
"in 5 secondi"

iex> Timex.Translator.translate_plural("invalid_locale", "relative_time", "in %{count} second", "in %{count} seconds", 5)
"in 5 seconds"

with_locale(locale, list) (macro)

This macro sets the locale during execution of a given block of code.